requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2
python-dotenv==1.0.0
pandas==2.1.4
schedule==1.2.1
requests==2.31.0
supabase==2.11.0
